prj/ppt/requirements.txt
gradio
requests
python-pptx
python-dotenv
Pillow
lxml
# dotenv
